Move CA-Pressure reduction to ca-calculators
