+++ src/app/calculator/utilities/utilities.module.ts
@@ -8,7 +8,6 @@ import { Co2SavingsModule } from './co2-savings/co2-savings.module';
 import { ElectricityReductionModule } from './electricity-reduction/electricity-reduction.module';
 import { NaturalGasReductionModule } from './natural-gas-reduction/natural-gas-reduction.module';
 import { WaterReductionModule } from './water-reduction/water-reduction.module';
-import { CompressedAirPressureReductionModule } from './compressed-air-pressure-reduction/compressed-air-pressure-reduction.module';
 import { SteamReductionModule } from './steam-reduction/steam-reduction.module';
 import { CashFlowModule } from './cash-flow/cash-flow.module';
 import { UnitConverterModule } from './unit-converter/unit-converter.module';
@@ -34,7 +33,6 @@ import { RouterModule } from '@angular/router';
         Co2SavingsModule,
         ElectricityReductionModule,
         NaturalGasReductionModule,
-        CompressedAirPressureReductionModule,
         SteamReductionModule,
         WaterReductionModule,
         PipeInsulationReductionModule,
