+++ src/app/dashboard/sidebar/calculator-list/calculator-list.component.html
@@ -85,6 +85,12 @@
   <span class="sidebar-item" [routerLink]="['/calculators/compressed-air-list']" routerLinkActive="selected">Compressed
     Air</span>
   <div class="directory-item">
+    <span [routerLink]="['/calculators/compressed-air-reduction']" routerLinkActive="selected"
+      class="sidebar-item">Compressed Air
+      Reduction</span>
+    <span [routerLink]="['/calculators/compressed-air-pressure-reduction']" routerLinkActive="selected"
+      class="sidebar-item">Compressed Air
+      Pressure Reduction</span>
     <span [routerLink]="['/calculators/bag-method']" routerLinkActive="selected" class="sidebar-item">Leak Loss
       Estimator - Bag Method</span>
     <span [routerLink]="['/calculators/pneumatic-air']" routerLinkActive="selected" class="sidebar-item">Pneumatic Air
@@ -130,9 +136,6 @@
       class="sidebar-item">Electricity Reduction</span>
     <span [routerLink]="['/calculators/natural-gas-reduction']" routerLinkActive="selected" class="sidebar-item">Natural
       Gas Reduction</span>
-    <span [routerLink]="['/calculators/compressed-air-pressure-reduction']" routerLinkActive="selected"
-      class="sidebar-item">Compressed Air
-      Pressure Reduction</span>
     <span [routerLink]="['/calculators/steam-reduction']" routerLinkActive="selected" class="sidebar-item">Steam
       Reduction</span>
     <span [routerLink]="['/calculators/water-reduction']" routerLinkActive="selected" class="sidebar-item">Water
