+++ src/app/treasure-hunt/calculators/calculators.module.ts
@@ -7,7 +7,6 @@ import { MotorDriveModule } from '../../calculator/motors/motor-drive/motor-driv
 import { NaturalGasReductionModule } from '../../calculator/utilities/natural-gas-reduction/natural-gas-reduction.module';
 import { ElectricityReductionModule } from '../../calculator/utilities/electricity-reduction/electricity-reduction.module';
 import { WaterReductionModule } from '../../calculator/utilities/water-reduction/water-reduction.module';
-import { CompressedAirPressureReductionModule } from '../../calculator/utilities/compressed-air-pressure-reduction/compressed-air-pressure-reduction.module';
 import { LightingReplacementModule } from '../../calculator/lighting/lighting-replacement/lighting-replacement.module';
 import { ModalModule } from 'ngx-bootstrap';
 
@@ -23,6 +22,7 @@ import { FormsModule } from '@angular/forms';
 import { SteamReductionModule } from '../../calculator/utilities/steam-reduction/steam-reduction.module';
 import { PipeInsulationReductionModule } from '../../calculator/utilities/pipe-insulation-reduction/pipe-insulation-reduction.module';
 import { CompressedAirReductionModule } from '../../calculator/compressed-air/compressed-air-reduction/compressed-air-reduction.module';
+import { CompressedAirPressureReductionModule } from '../../calculator/compressed-air/compressed-air-pressure-reduction/compressed-air-pressure-reduction.module';
 
 @NgModule({
   declarations: [
